Rename namespace engine to namespace openssl_interface
